app/requirements.txt
aniso8601==8.0.0
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-RESTful==0.3.9
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pytz==2020.1
six==1.15.0
Werkzeug==2.0.2
gunicorn==20.1.0
mysql-connector-python
